Fix analyzer worker metric assignment
The existing formula for computing the fraction of the metrics to
process for each worker subprocess could result in situations where the
jobs of the last worker overlapped with the ones of the previous worker
or even worse, resulted in an index exception. This commit corrects this
issue.

To visualize the existing issue: With the existing calculations, having
11 metrics and 5 workers resulted in the following provisioning
scheme:
```
1: [0, 1, 2]
2: [3, 4, 5]
3: [6, 7, 8]
4: [9, 10, 11]
5: [8, 9, 10]
```
Worker 4 creates an index error and worker 5 overlaps.